requirements-strategy-testing.txt
# Trading Strategy Testing Framework Dependencies
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data sources
alpaca-trade-api>=3.0.0
alpaca-py>=0.20.0
polygon-api-client>=1.12.0
yfinance>=0.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistical analysis
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Async and HTTP
aiohttp>=3.8.0
requests>=2.31.0

# Data validation
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
diskcache>=5.6.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
